packages/claude_validation/pyproject.toml
[project]
name = "pnseg-claude_validation"
version = "0.1.0"
dependencies = [
    "pnseg-core",
    "anthropic",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["claude_validation*"]